build: support --mode={config|build} for custom cli args
Before the mpv-build rewrite, scripts/${proj}-{config|build} were stand
alone and used hardcoded projects source paths, therefore it was
possible to run e.g. ./scripts/mpv-config <config-args> &&
./scripts/mpv-build <build-args>

However, now that these scripts depend on config_${prog}_source
they're no longer stand alone. Additionally, the high-level ./build
script didn't allow an equivalent procedure because it always invoked
both configure and build, but passed the additional cli arguments only
to build (because it's impractical to have the same invocation with both
config and build arguments).

We can solve it in one of two ways:
- Make those scripts independent by having them read the config files too.
- Enhance the ./build script to support such procedure.

The former is problematic because this means every new project we add
will need to read the config files, and additionally it will have to
know not to read them if invoked by ./build. It's a bit of a mess.

The other approach actually turns out simple and elegant: allow build to
run like now (configure+build, args go to build), or do configure-only
(args go to configure) or build-only (args go to build), the last two
allow an equivalent procedure.

However, such procedure is a bit of a footgun, because running configure
on few projects and then build on those projects is not the same as
running configure+build on each project. e.g. mpv-config actually needs
that ffmpeg-build was performed and will not be satisfied by only
ffmpeg-config.

So for now we don't advertise these modes, but now ./build also
optionally accepts, only as its first argument: --mode=config or
--mode=build which perform only configure/build and each accepts custom
cli args for its own procedure.