rewrite of install/uninstall - separate process from data
install and uninstall were hardcoded to run scripts/mpv-[un]install,
however this is data (mpv) and we want to avoid it at the global
scripts, so we we do the same as for the debian changelog generator and
consider the last item of $config_projects as the main project which
should be [un]installed.

The user can override $config_projects from CLI, but it's not guaranteed
than an install script will exist for projects other than the main one.

We also pass additional arguments, which we previously didn't do.