hooks: don't use subshell
There's no real reason to prevent the user from running hooks in the
host shell, and nothing prevents the user, if they want to, to execute
things in a subshell.

However, forcing a subshell limits the value of hooks as they cannot
affect the host anymore, for inctance by modifying CFLAGS etc, which
the next commit will use.