projects: support shaderc and crossc (disabled by default)
Both are c++ libs and both have dependency repositories.

crossc has submodules, and while in the future we might support it as
standard, for now we use an update-post hook which does the submodules
thingy. Its .pc file has -lstdc++ at the Libs section if built as
static, so mpv doesn't need to probe it with --static .

shaderc uses several repositories and has an update-post hook which
updates them always to latest. Additionally it doesn't have a .pc file
so we create one at shaderc-build . Like crossc, we put -lstdc++ at the
Libs section and expose only the static libs, so mpv's wscript can do
plain check_pkg_config and work with it as is without static concerns.