ffmpeg: use build dir inside prefix/tmp
Previously it was creating an ffmpeg_build at the root dir - which was
polluting it. We could have created the build dir inside the source dir,
but it creates a symlink to the source dir inside it, and while it does
work on windows, it's dodgy because in msys2 symlink is plain copy,
which has to avoid recursion.

We could also have done the build plainly at the source dir, but a build
dir is cleaner, and prefix/tmp is just fine for it.