config: allow config_local_prefix at user.conf, with some protection
Previously the prefix could only be configured at config/default.conf
because clean can delete it with bad consequences if it points to some
place which shouldn't be deleted.

Now it can be configured, but we add some protections:
- When we create the prefix dir, we also touch "core-prefix" in it.
- clean checks for <prefix>/core-prefix before deleting the prefix.

This will prevent clean after updating mpv-build, but it's transitional
and the message states what needs to be done (touch <prefix>/core-prefix).