update: don't use shallow with an existing full repo
That is, --depth=... will not be used if the repo dir exists and is not
a shallow repo. This is mostly to ease working with existing git repos
without turning them into shallow.

The user will have to manually change shallowness, or delete the source
dir and update it from scratch - which will respect the shallow config.