config: move foo_options to config_foo_opts_{config|build} at user.conf
foo_options is now ignored, and a warning is displayed if it exists.

For both config values, the arguments are centrally-prepended to the final
CLI arguments during configure/build respectively, and passed to
scripts/foo-{config|build} - instead of those scripts doing the
concatenation themselves (with duplicated code).

The config values are space-separated strings, and currently space
in individual arguments is not supported - like before, but this might
change in the future (TBD: how? line per arg? backslash-space?).

Because mpv_options doesn't exist anymore and scripts/test-libmpv was
using it, now scripts/core-config exports the main project opts_config,
and test-libmpv tests config_mpv_opts_config instead.

Note that if --enable-libmpv (which test-libmpv checks) is not used at the
config file and is used via cli then it cannot detect it - same as before.

Additionally, make sure all the foo-{config|build} scripts now use the
input arguments, and fix some $@ to "$@".