PIC: support remaining  projects, disable for mingw
Only mujs actually supports it, where crossc doesn't (but doesn't need
to - windows), and shaderc apparently always enables it (regardless, not
sure how to request it to use -fPIC).

But add the handler, even if empty, to the remaining projects, where
ffmpeg/libass/fribidi already supported it.

Additionally, auto disable PIC for windows (if not set explicitly to
yes/no) because apparently it's not required.