tools/auto_ccmpp: fallback: just our prefix and only the flags
Previously the fallback behavior (no args and not in msys2/osx) was to
deprioritize /usr/local. Now we don't deprioritize any specific dir and
just add our prefix to the flags, and don't touch pkg-config env vars.

This is a more conservative default behavior - which also works with a
cross-build setup where /usr/local is not necessarily host-specific.

The build system typically uses the CFLAGS before other -I... so it
should still be useful generally.