ffmpeg config: detect mbedtls/schannel/securetransport options
We still don't auto-enable those because its not a simple pkg-config
and we don't want to dwelve into more complex detections, but now we
at least skip our auto-enable if we notice one of them at the options.

Also, previously we skipped auto-enable if one of the ssl/tls options
was enabled at the options. Now we skip it if one of them is mentioned
at all (either enabled or disabled) - better to skip our meddling as
soon as we notice the user does related things manualy.