Use python3 for /usr/bin/python if python2 is no longer available
As distros start to remove python2 from their repositories, we need to make
sure waf can still run python from its shebang.

The order of the alternatives is chosen such that distros that have the
python3-docutils package but not python-is-python3 will install a
compatible combination (namely, python and python-docutils).