libass: static linking: use statify-pc.sh
The previous hack was added at 85d631f . This commit replaces it.

While both the previous solution and the current solution are hacks
to allow static linking, the previous solution was too hacky:
- Hardcoded list of deps which should be linked with libass.
- ... added one-by-one because different platforms are different.
- Added junk to the LDFLAGS env, which affects everything, and was
  making ./waf configure very slow for MINGW.

The new hack is a generic .pc modifier which does the "right" thing
(relatively speaking) automatically and correctly, without polluting
LDFLAGS. Also, it backs-up the original .pc file as libass.pc.bak.