rebuild: only delete prefix when invoked without arguments
rebuild is a shortcut to clean+update+build, and until now it also
unconditionally deleted the (mpv-build) prefix and all the installed
libs in it.

This behavior is fine when rebuilding all the packages, but
unexpected when one rebuilds only a subset of the packages, like
with `./rebuild ffmpeg mpv` .

Now the prefix is deleted only when rebuild is invoked without
arguments (where it rebuilds all the configured packages).