lua: switch repo to https://github.com/avih/lua-cmake
This is a fork of https://github.com/LuaDist/lua because the
original is no longer maintained (marked as "archived" on github).

Of mainline lua (not jit) mpv only supports lua 5.1 and 5.2, which
are not maintained anymore upstream, therefore further commits are
not expected, and the repo already has the latest/last mainline
versions (5.1.5 and 5.2.4).

However, published fixes, and/or cmake build adjustments may still
happen.

This fork now already includes the latest known 5.1 fixes (one CVE,
one minor bug), as well as a cmake-build fix at the lua-5.1 branch
to respect BUILD_SHARED_LIBS=OFF .

The branches lua-5.1 and lua-5.2 are the latest versions respectively.
Only the tip of these branches is recommended, as other tags/branches
don't include all the fixes.

We use a cmake build repo and not the upstream/mainline lua repo
because the mainline makefile is not great.