lua: split into lua51 and lua52 projects
Lua 5.1 and 5.2 are typically considered separate projects by distros
and package managers, so we do the same, even if it's the same repo.

Makes it easier to build one or the other. Also they now have separate
pkg-config file names (lua5X.pc) and separate lib names (liblua5X.a).