posix sh: don't use local vars (almost)
In many of the places local wasn't actually needed (not called from
places which use the same var name), or, where it has a more lasting
effect or using more common names (short) - a unique prefix is used.

While it's possible to add a "system" for local variables, for now
it's not yet needed, and per-case solutions are enough - with care.

The only remaining use of local is at scripts/mpv-conf-utils.sh,
and only at cross-compile setup code (host=...). for now it remains.

We _should_ be posix sh compliant now, except when cross compiling.
We do expect a standard/unset IFS, but it becomes standard on startup
(by posix) and it's not modified by our scripts, so that's fine.

Obviously, we don't control 3rd party build scripts, like "configure".