build: respect "set -e" in pkg recipes (lua*, hb, zimg, sixel)
The sh form "foo || bar" disables/prevents "set -e" in foo except
when foo is an external script/command. I.e. if foo is a shell
function or shell code (list or subshell) then even if it does use
"set -e" (like all pkg recipes do) then it's ignored.

The build script used such form when running the pkg recipes,
therefore their "set -e" was not respected (all of them set it).

Now it's respected.

This is a change in behavior due to this fix, but for the better.