mbedtls: restore to LTS branch 2.16.<LATEST>
We can detect the latest sub-version (latest security fixes) of
the 2.16 LTS branch, but with the 2.x (and the recent 3.x), the
branch versions mix release and development tags, and therefore
harder to detect a latest release of this branch with all security
fixes but excluding development snapshots. Stick to 2.16.* for now.