update: fix grep usage when specifying a pull-request
./update foo/^PR1234 (and ^#1234 and ^1234) are supposed to update
project foo to pull-request 1234, however, grep was used incorrectly,
which resulted in this not working on some systems (e.g. illumos).

Specifically, ERE alternation (foo|bar) is not supposed to work with
BRE (default for grep - which we used), but we used a gnu-specific
form foo\|bar which doesn't work with all grep implementation.

Now we use ERE correctly (and escaping the parenthesis is not needed).

Note: we didn't, and still don't try to escape the user value, though
while possible, it's unlikely to result in incorrect behavior.